+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* **Breaking**: `cmds` prop are no longer supported; use the imperative API instead (e.g. instead of passing down a `REVERT` command in `cmds`, call `revert()` on the corresponding input). See the docs for more details.
 * **Breaking**: We no longer set input `id`/`name` on many components by default (e.g. checkboxes, text/password/number inputs, selects, etc.). If you use input labels or radio groups, make sure you explicitly set the input's `id` prop.
 * DataTable, ColorInput, DateInput: **improved behaviour on touch-based devices**.
 * Add iOS-specific styles in the Giu component – **if you target the iOS platform, please include this component above all other Giu components**.